Add pristine gl2ps.h gl2ps.c files

+++ lib/gl2ps.c
@@ -186,7 +186,7 @@ typedef struct{
   int gsobjno, fontobjno, imobjno, shobjno, maskshobjno, trgroupobjno;
 } GL2PSpdfgroup;
 
-typedef struct { 
+typedef struct {
   /* General */
   GLint format, sort, options, colorsize, colormode, buffersize;
   GLint lastlinecap, lastlinejoin;
@@ -4433,7 +4433,7 @@ static int gl2psPrintPDFShaderStreamDataCoord(GL2PSvertex *vertex,
   int offs = 0;
   unsigned long imap;
   GLfloat diff;
-  double dmax = (double) ~1UL;
+  double dmax = ~1UL;
   char edgeflag = 0;
 
   /* FIXME: temp bux fix for 64 bit archs: */
@@ -4476,7 +4476,7 @@ static int gl2psPrintPDFShaderStreamDataRGB(GL2PSvertex *vertex,
 {
   int offs = 0;
   unsigned long imap;
-  double dmax = (double) ~1UL;
+  double dmax = ~1UL;
   GLfloat tmp;
 
   /* FIXME: temp bux fix for 64 bit archs: */
@@ -4517,7 +4517,7 @@ static int gl2psPrintPDFShaderStreamDataAlpha(GL2PSvertex *vertex,
 {
   int offs = 0;
   unsigned long imap;
-  double dmax = (double) ~1UL;
+  double dmax = ~1UL;
   GLfloat tmp;
 
   /* FIXME: temp bux fix for 64 bit archs: */
